+++ tests/integration_test.rs
@@ -14,50 +14,15 @@ fn fresh(){
 
 #[serial]
 #[test]
-fn test_profile_vs_query(){
-
-    let mut output = Command::cargo_bin("sylph").unwrap();
-    let output = output
-        .arg("profile")
-        .arg("./test_files/o157_reads.fastq")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .output()
-        .expect("Output failed");
-    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
-    dbg!(stdout.matches('\n').count());
-    assert!(stdout.matches('\n').count() == 2);
-
-    let mut output = Command::cargo_bin("sylph").unwrap();
-    let output = output
-        .arg("query")
-        .arg("./test_files/o157_reads.fastq")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .arg("./test_files/e.coli-o157.fasta")
-        .arg("./test_files/e.coli-K12.fasta")
-        .output()
-        .expect("Output failed");
-    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
-    dbg!(stdout.matches('\n').count());
-    println!("{}",stdout);
-    assert!(stdout.matches('\n').count() == 4);
-}
-
-#[serial]
-#[test]
 fn test_sketch_commands() {
-    Command::new("rm")
-        .arg("-r")
-        .args(["./tests/results/test_sketch_dir"])
-        .spawn();
-    let mut cmd = Command::cargo_bin("sylph").unwrap();
+   let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("sketch")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .arg("./test_files/e.coli-K12.fasta")
-        .arg("./test_files/o157_reads.fastq")
-        .arg("./test_files/e.coli-W.fasta.gz")
+        .arg("test_files/e.coli-EC590.fasta.gz")
+        .arg("test_files/e.coli-K12.fasta.gz")
+        .arg("test_files/o157_reads.fastq.gz")
         .arg("-o")
-        .arg("./tests/results/test_sketch_dir/db")
+        .arg("tests/results/test_sketch_dir/db")
         .arg("-d")
         .arg("./tests/results/test_sketch_dir")
         .assert();
@@ -66,7 +31,7 @@ fn test_sketch_commands() {
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("profile")
-        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp")
         .arg("./tests/results/test_sketch_dir/db.syldb")
         .assert();
     assert.success().code(0);
@@ -83,16 +48,16 @@ fn test_sketch_commands() {
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("profile")
-        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
-        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
         .assert();
     assert.success().code(0);
 
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("profile")
-        .arg("./test_files/o157_reads.fastq")
-        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/o157_reads.fastq.gz")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
         .arg("-i")
         .arg("-m")
         .arg("90")
@@ -126,8 +91,6 @@ fn test_sketch_commands() {
     assert.success().code(0);
     assert!(Path::new("./tests/results/test_sketch_dir/t1.fq.paired.sylsp").exists(), "Output file was not created");
 
-
-
     fresh();
     let mut cmd= Command::cargo_bin("sylph").unwrap();
     let assert = cmd
@@ -144,40 +107,74 @@ fn test_sketch_commands() {
     assert.success().code(0);
     assert!(Path::new("./tests/results/test_sketch_dir/t2.fq.sylsp").exists(), "Output file was not created");
     assert!(Path::new("./tests/results/test_sketch_dir/testdb.syldb").exists(), "Output file was not created");
+}
 
+#[serial]
+#[test]
+fn test_profile_vs_query(){
+    fresh();
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq.gz")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .output()
+        .expect("Output failed");
+    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+    dbg!(stdout.matches('\n').count());
+    assert!(stdout.matches('\n').count() == 2);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("query")
+        .arg("./test_files/o157_reads.fastq.gz")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("./test_files/e.coli-o157.fasta.gz")
+        .arg("./test_files/e.coli-K12.fasta.gz")
+        .output()
+        .expect("Output failed");
+    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+    dbg!(stdout.matches('\n').count());
+    println!("{}",stdout);
+    assert!(stdout.matches('\n').count() == 4);
+}
+
+#[serial]
+#[test]
+fn test_sketch_list(){
     fresh();
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("sketch")
         .arg("-r")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("-o")
         .arg("./tests/results/test_sketch_dir/db")
         .arg("-d")
         .arg("./tests/results/test_sketch_dir")
         .assert();
     assert.success().code(0);
-    assert!(Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.sylsp").exists(), "Output file was not created");
-    assert!(Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp").exists(), "Output file was not created");
+    assert!(Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.gz.sylsp").exists(), "Output file was not created");
+    assert!(Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp").exists(), "Output file was not created");
     assert!(!Path::new("./tests/results/test_sketch_dir/db.syldb").exists(), "Output file was created");
     fresh();
 
-    fresh();
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("sketch")
         .arg("-g")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("-o")
         .arg("./tests/results/test_sketch_dir/db")
         .arg("-d")
         .arg("./tests/results/test_sketch_dir")
         .assert();
     assert.success().code(0);
-    assert!(!Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.sylsp").exists(), "Output file was created");
-    assert!(!Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp").exists(), "Output file was created");
+    assert!(!Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.gz.sylsp").exists(), "Output file was created");
+    assert!(!Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp").exists(), "Output file was created");
     assert!(Path::new("./tests/results/test_sketch_dir/db.syldb").exists(), "Output file was not created");
     fresh();
 
@@ -205,12 +202,11 @@ fn test_sketch_commands() {
         .assert();
     assert.success().code(0);
     assert!(!Path::new("./tests/results/test_sketch_dir/db.syldb").exists(), "Output file was not created");
-    assert!(Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.sylsp").exists(), "Output file was not created");
-    assert!(Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp").exists(), "Output file was not created");
+    assert!(Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.gz.sylsp").exists(), "Output file was not created");
+    assert!(Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp").exists(), "Output file was not created");
     fresh();
 
 }
-
 #[serial]
 #[test]
 fn test_profile_disabling(){
@@ -220,7 +216,7 @@ fn test_profile_disabling(){
     let assert = cmd
         .arg("sketch")
         .arg("-g")
-        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
         .arg("-o")
         .arg("./tests/results/test_sketch_dir/db")
         .arg("-d")
@@ -232,7 +228,7 @@ fn test_profile_disabling(){
     let mut output = Command::cargo_bin("sylph").unwrap();
     let assert = output
         .arg("profile")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("./tests/results/test_sketch_dir/db.syldb")
         .assert();
     assert.failure().code(1);
@@ -240,14 +236,13 @@ fn test_profile_disabling(){
     let mut output = Command::cargo_bin("sylph").unwrap();
     let assert = output
         .arg("query")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("./tests/results/test_sketch_dir/db.syldb")
         .assert();
     assert.success().code(0);
 
     fresh();
 }
-
 #[serial]
 #[test]
 fn test_sketch_fasta_fastq_concord(){
@@ -255,8 +250,8 @@ fn test_sketch_fasta_fastq_concord(){
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("sketch")
-        .arg("./test_files/e.coli-EC590.fasta")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("-o")
         .arg("./tests/results/test_sketch_dir/db")
         .arg("-d")
@@ -267,7 +262,7 @@ fn test_sketch_fasta_fastq_concord(){
     let mut output = Command::cargo_bin("sylph").unwrap();
     let out1 = output
         .arg("profile")
-        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/o157_reads.fastq.gz")
         .arg("./tests/results/test_sketch_dir/db.syldb")
         .output()
         .expect("Fail");
@@ -275,15 +270,15 @@ fn test_sketch_fasta_fastq_concord(){
     let mut output = Command::cargo_bin("sylph").unwrap();
     let out2 = output
         .arg("profile")
-        .arg("./test_files/o157_reads.fastq")
-        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/o157_reads.fastq.gz")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
         .output()
         .expect("Fail");
 
     let mut output = Command::cargo_bin("sylph").unwrap();
     let out3 = output
         .arg("profile")
-        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.gz.sylsp")
         .arg("./tests/results/test_sketch_dir/db.syldb")
         .output()
         .expect("Fail");
@@ -298,10 +293,10 @@ fn test_sketch_fasta_fastq_concord(){
 
     fresh();
 }
-
 #[serial]
 #[test]
 fn test_sample_names(){
+    fresh();
     let mut cmd = Command::cargo_bin("sylph").unwrap();
     let assert = cmd
         .arg("sketch")
@@ -322,7 +317,7 @@ fn test_sample_names(){
     let assert = cmd
         .arg("sketch")
         .arg("test_files/t1.fq")
-        .arg("test_files/o157_reads.fastq")
+        .arg("test_files/o157_reads.fastq.gz")
         .arg("-d")
         .arg("./tests/results/test_sketch_dir")
         .arg("--lS")
@@ -336,7 +331,7 @@ fn test_sample_names(){
     let output = cmd
         .arg("profile")
         .arg("./tests/results/test_sketch_dir/S2.sylsp")
-        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
         .output().unwrap();
     let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
     dbg!(&stdout);
@@ -377,10 +372,7 @@ fn test_sample_names(){
     assert!(Path::new("./tests/results/test_sketch_dir/SAMPLE_TEST_S1.paired.sylsp").exists(), "Output file was not created, -S");
 
     fresh();
-
-
 }
-
 #[serial]
 #[test]
 fn test_fpr(){
@@ -427,4 +419,83 @@ fn test_fpr(){
     assert.failure().code(1);
     fresh();
 
+}
+#[serial]
+#[test]
+fn test_raw_inputs_profile_simple(){
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./test_files/e.coli-o157.fasta.gz")
+        .arg("-1")
+        .arg("test_files/t1.fq")
+        .arg("-2")
+        .arg("test_files/t2.fq")
+        .assert();
+    assert.success().code(0);
+    fresh();
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./test_files/e.coli-o157.fasta.gz")
+        .arg("-1")
+        .arg("test_files/t1.fq")
+        .assert();
+    assert.failure().code(1);
+    fresh();
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./test_files/e.coli-o157.fasta.gz")
+        .arg("-1")
+        .arg("test_files/k12_R1.fq")
+        .arg("test_files/t1.fq")
+        .arg("-2")
+        .arg("test_files/k12_R2.fq")
+        .arg("test_files/t1.fq")
+        .assert();
+    assert.success().code(0);
+    fresh();
+    
+}
+#[serial]
+#[test]
+fn test_raw_inputs_profile_with_sketch(){
+    
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("profile")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("-1")
+        .arg("test_files/k12_R1.fq")
+        .arg("-2")
+        .arg("test_files/k12_R2.fq")
+        .output()
+        .expect("Output failed");
+    let stdout_1 = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("-1")
+        .arg("test_files/k12_R1.fq")
+        .arg("-2")
+        .arg("test_files/k12_R2.fq")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("profile")
+        .arg("./test_files/e.coli-EC590.fasta.gz")
+        .arg("./tests/results/test_sketch_dir/k12_R1.fq.paired.sylsp")
+        .output()
+        .expect("Output failed");
+    let stdout_2 = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+
+    assert!(stdout_1 == stdout_2);
 }